requirements.txt
altair==5.2.0
anyio==4.2.0
appdirs==1.4.4
arrow==1.3.0
attrs==23.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
blinker==1.7.0
cachetools==5.3.2
certifi==2021.10.8
chardet==5.2.0
charset-normalizer==2.0.10
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
cycler==0.11.0
DateTime==4.3
exceptiongroup==1.2.0
fix-yahoo-finance==0.1.37
flet==0.19.0
flet-core==0.19.0
flet-runtime==0.19.0
fonttools==4.29.0
frozendict==2.3.10
gitdb==4.0.11
GitPython==3.1.41
greenlet==1.1.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.3
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.3.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.5.1
mdurl==0.1.2
mplfinance==0.12.8b6
multitasking==0.0.10
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==1.4.0
pandas-datareader==0.10.0
patsy==0.5.6
peewee==3.17.0
Pillow==9.0.0
pip-review==1.3.0
plotly==5.18.0
plotly-express==0.4.1
protobuf==4.25.2
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.0.7
pypng==0.20220715.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0
qrcode==7.4.2
referencing==0.33.0
repath==0.9.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.31
statsmodels==0.14.1
streamlit==1.31.0
streamlit-autorefresh==1.0.1
TA-Lib @ file:///C:/TA_Lib-0.4.24-cp310-cp310-win_amd64.whl
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.8
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
xlrd==2.0.1
yahoo-finance-api2==0.0.12
yfinance==0.2.33
zipp==3.17.0
zope.interface==5.4.0
